+++ test/unit/spec/signaling/v2/localparticipant.js
@@ -56,6 +56,14 @@ describe('LocalParticipantV2', () => {
   });
 
   describe('constructor', () => {
+    it('should set .bandwidthProfile to null', () => {
+      assert.equal(localParticipant.bandwidthProfile, null);
+    });
+
+    it('should set .bandwidthProfileRevision to 0', () => {
+      assert.equal(localParticipant.bandwidthProfileRevision, 0);
+    });
+
     it('should set .networkQualityConfiguration', () => {
       assert.equal(localParticipant.networkQualityConfiguration, networkQualityConfiguration);
     });
@@ -191,6 +199,95 @@ describe('LocalParticipantV2', () => {
           localParticipant.once(event, () => events.push(event)));
         localParticipant.removeTrack(trackSender);
         assert.deepEqual(events, []);
+      });
+    });
+  });
+
+  describe('#setBandwidthProfile', () => {
+    [
+      ['dominantSpeakerPriority', 'high'],
+      ['maxSubscriptionBitrate', 3000],
+      ['maxTracks', 4],
+      ['mode', 'grid'],
+      ['renderDimensions.high', { width: 4 }],
+      ['renderDimensions.low', { height: 2 }],
+      ['renderDimensions.standard', { width: 3 }],
+      ['']
+    ].forEach(([prop, value]) => {
+      context(`when called with a BandwidthProfileOptions that ${prop ? `has a different .${prop}` : 'is the same'}`, () => {
+        let bandwidthProfileRevision;
+        let initialBandwidthProfile;
+        let newBandwidthProfile;
+        let revision;
+        let updated;
+
+        beforeEach(() => {
+          initialBandwidthProfile = {
+            video: {
+              dominantSpeakerPriority: 'low',
+              maxSubscriptionBitrate: 2000,
+              maxTracks: 3,
+              mode: 'collaboration',
+              renderDimensions: {
+                high: { height: 3, width: 3 },
+                low: { height: 1, width: 1 },
+                standard: { height: 2, width: 2 }
+              }
+            }
+          };
+
+          localParticipant.setBandwidthProfile(initialBandwidthProfile);
+          bandwidthProfileRevision = localParticipant.bandwidthProfileRevision;
+          revision = localParticipant.revision;
+          updated = false;
+
+          const [mainProp, subProp] = prop.split('.');
+          newBandwidthProfile = JSON.parse(JSON.stringify(initialBandwidthProfile));
+          if (subProp) {
+            newBandwidthProfile.video[mainProp][subProp] = Object.assign(newBandwidthProfile.video[mainProp][subProp], value);
+          } else if (mainProp) {
+            newBandwidthProfile.video[mainProp] = value;
+          }
+          localParticipant.once('updated', () => { updated = true; });
+
+          localParticipant.setBandwidthProfile(newBandwidthProfile);
+        });
+
+        if (prop) {
+          it('should set .bandwidthProfile to the new BandwidthProfileOptions', () => {
+            assert.deepEqual(localParticipant.bandwidthProfile, newBandwidthProfile);
+          });
+
+          it('should increment .bandwidthProfileRevision', () => {
+            assert.equal(localParticipant.bandwidthProfileRevision, bandwidthProfileRevision + 1);
+          });
+
+          it('should increment .revision', () => {
+            assert.equal(localParticipant.revision, revision + 1);
+          });
+
+          it('should emit "updated"', () => {
+            assert(updated);
+          });
+
+          return;
+        }
+
+        it('should not change .bandwidthProfile', () => {
+          assert.deepEqual(localParticipant.bandwidthProfile, initialBandwidthProfile);
+        });
+
+        it('should not increment .bandwidthProfileRevision', () => {
+          assert.equal(localParticipant.bandwidthProfileRevision, bandwidthProfileRevision);
+        });
+
+        it('should not increment .revision', () => {
+          assert.equal(localParticipant.revision, revision);
+        });
+
+        it('should not emit "updated"', () => {
+          assert(!updated);
+        });
       });
     });
   });